Adds Animation Frame De-Duplication (#2)
* Adds Animation Frame De-Duplication

One of the big heartaches for wallening (among many) is animated wall
icons
When they're too large, we get massive lag when going near a wall. This
is particularly bad when there's a lot of frames
There are some cases where they could be split off into overlays,
applied seperately. This doesn't apply to all cases though

So I was thinking, there's surely a lot of duplicate frames here, at
least for the major suck (plasma walls, currently 2.7 mbs)
Action is busy with uh, I think a UI, so I decided to take a crack at it
myself

I've written a function that walks the frames of an icon state, and
simply removes the duplicates, extending the animation time
This isn't the fastest possible implementation, both because this is my
first time writing rust, and because I'm not like, doing it before
insertion

It's more general tho, and lets us apply the proc to all the uses of
Icon State in the codebase

Sorry if uh, it's crap. Doin what I can

Oh right nearly forgot, this drops plasma walls down more then a
megabyte, to 1.2mbs. Doesn't really do anything to heirophont walls, the
other use of animations we have, but that's alright, I can make those
into overlays anyhow

* Fuck, ensures delay index always increments

* removes seemingly unneeded clone()s (I need to read on these more)

* formatting pass

* Converts deduping to fold() rather then a for loop

Easier to read.
Also uh, I never neaded to clone() the frame list, since uh, frames
implement partialEq, and since they're just lists of bytes it's
basically free.

* trailing return

---------

Co-authored-by: Rob Bailey <actioninja@criticalaction.net>